Upgrade babel and bring babel runtime back (#73)
* Upgrade babel and bring babel runtime back

The problem you achieved was because of wrapped with runtime cjs helpers
and polyfills. In this diff I upgraded babel and added esm versions of
helpers. Polyfills are completely disabled which is the same behaviour
as inlined helpers. The size is increased because helpers in babel7 was
a bit changed.

* let's merge it