Transpile in loose mode and specify correct babel runtime (#76)
* Transpile in loose mode and specify correct babel runtime

Loose mode works well here. Small size improvement.

About babel runtime. It should be in sync with others versions because
with every next version it may be changed. You can't be sure even these
versions do not cause bugs just because it works right now.

Syncing version with mui doesn't make sense. Then mui will be updated
and won't be in sync with this project. It's better to update whenever
possible to the latest versions.

There are different problems which we may solve. Do not stuck on helpers
please. You won't win a lot here. They will be 100% fixed when babel 7 will be
relased.

* let's be safe